# MAINT: Optimize the performance of count_nonzero by using universal intrinsics