--- a/numpy/core/src/common/simd/avx2/conversion.h
+++ b/numpy/core/src/common/simd/avx2/conversion.h
@@ -29,4 +29,17 @@
 #define npyv_cvt_b32_f32(BL) _mm256_castps_si256(BL)
 #define npyv_cvt_b64_f64(BL) _mm256_castpd_si256(BL)
 
+// expand
+NPY_FINLINE void npyv_expand_u8_u16(npyv_u8 data, npyv_u16 *low, npyv_u16 *high) {
+    const __m256i z = _mm256_setzero_si256();
+    *low = _mm256_unpacklo_epi8(data, z);
+    *high = _mm256_unpackhi_epi8(data, z);
+}
+
+NPY_FINLINE void npyv_expand_u16_u32(npyv_u16 data, npyv_u32 *low, npyv_u32 *high) {
+    const __m256i z = _mm256_setzero_si256();
+    *low = _mm256_unpacklo_epi16(data, z);
+    *high = _mm256_unpackhi_epi16(data, z);
+}
+
 #endif // _NPY_SIMD_AVX2_CVT_H