--- a/numpy/core/src/common/simd/avx2/conversion.h
+++ b/numpy/core/src/common/simd/avx2/conversion.h
@@ -30,16 +30,18 @@
 #define npyv_cvt_b64_f64(BL) _mm256_castpd_si256(BL)
 
 // expand
-NPY_FINLINE void npyv_expand_u8_u16(npyv_u8 data, npyv_u16 *low, npyv_u16 *high) {
-    const __m256i z = _mm256_setzero_si256();
-    *low = _mm256_unpacklo_epi8(data, z);
-    *high = _mm256_unpackhi_epi8(data, z);
+NPY_FINLINE npyv_u16x2 npyv_expand_u8_u16(npyv_u8 data) {
+    npyv_u16x2 r;
+    r.val[0] = _mm256_cvtepu8_epi16(_mm256_castsi256_si128(data));
+    r.val[1] = _mm256_cvtepu8_epi16(_mm256_extracti128_si256(data, 1));
+    return r;
 }
 
-NPY_FINLINE void npyv_expand_u16_u32(npyv_u16 data, npyv_u32 *low, npyv_u32 *high) {
-    const __m256i z = _mm256_setzero_si256();
-    *low = _mm256_unpacklo_epi16(data, z);
-    *high = _mm256_unpackhi_epi16(data, z);
+NPY_FINLINE npyv_u32x2 npyv_expand_u16_u32(npyv_u16 data) {
+    npyv_u32x2 r;
+    r.val[0] = _mm256_cvtepu16_epi32(_mm256_castsi256_si128(data));
+    r.val[1] = _mm256_cvtepu16_epi32(_mm256_extracti128_si256(data, 1));
+    return r;
 }
 
 #endif // _NPY_SIMD_AVX2_CVT_H