--- a/numpy/core/src/common/simd/avx2/conversion.h
+++ b/numpy/core/src/common/simd/avx2/conversion.h
@@ -30,14 +30,14 @@
 #define npyv_cvt_b64_f64(BL) _mm256_castpd_si256(BL)
 
 // expand
-NPY_FINLINE npyv_u16x2 npyv_expand_u8_u16(npyv_u8 data) {
+NPY_FINLINE npyv_u16x2 npyv_expand_u16_u8(npyv_u8 data) {
     npyv_u16x2 r;
     r.val[0] = _mm256_cvtepu8_epi16(_mm256_castsi256_si128(data));
     r.val[1] = _mm256_cvtepu8_epi16(_mm256_extracti128_si256(data, 1));
     return r;
 }
 
-NPY_FINLINE npyv_u32x2 npyv_expand_u16_u32(npyv_u16 data) {
+NPY_FINLINE npyv_u32x2 npyv_expand_u32_u16(npyv_u16 data) {
     npyv_u32x2 r;
     r.val[0] = _mm256_cvtepu16_epi32(_mm256_castsi256_si128(data));
     r.val[1] = _mm256_cvtepu16_epi32(_mm256_extracti128_si256(data, 1));