--- a/numpy/core/src/common/simd/avx512/conversion.h
+++ b/numpy/core/src/common/simd/avx512/conversion.h
@@ -52,16 +52,18 @@
 #define npyv_cvt_b64_f64(A) npyv_cvt_b64_u64(_mm512_castpd_si512(A))
 
 // expand
-NPY_FINLINE void npyv_expand_u8_u16(npyv_u8 data, npyv_u16 *low, npyv_u16 *high) {
-    const __m512i z = _mm512_setzero_si512();
-    *low = _mm512_unpacklo_epi8(data, z);
-    *high = _mm512_unpackhi_epi8(data, z);
+NPY_FINLINE npyv_u16x2 npyv_expand_u8_u16(npyv_u8 data) {
+    npyv_u16x2 r;
+    r.val[0] = _mm512_cvtepu8_epi16(_mm512_castsi512_si256(data));
+    r.val[1] = _mm512_cvtepu8_epi16(_mm512_extracti32x8_epi32(data, 1));
+    return r;
 }
 
-NPY_FINLINE void npyv_expand_u16_u32(npyv_u16 data, npyv_u32 *low, npyv_u32 *high) {
-    const __m512i z = _mm512_setzero_si512();
-    *low = _mm512_unpacklo_epi16(data, z);
-    *high = _mm512_unpackhi_epi16(data, z);
+NPY_FINLINE npyv_u32x2 npyv_expand_u16_u32(npyv_u16 data) {
+    npyv_u16x2 r;
+    r.val[0] = _mm512_cvtepu16_epi32(_mm512_castsi512_si256(data));
+    r.val[1] = _mm512_cvtepu16_epi32(_mm512_extracti32x8_epi32(data, 1));
+    return r;
 }
 
 #endif // _NPY_SIMD_AVX512_CVT_H