--- a/numpy/core/src/common/simd/avx512/conversion.h
+++ b/numpy/core/src/common/simd/avx512/conversion.h
@@ -55,13 +55,18 @@
 NPY_FINLINE npyv_u16x2 npyv_expand_u16_u8(npyv_u8 data)
 {
     npyv_u16x2 r;
+    __m256i lo = _mm512_castsi512_si256(data);
+    __m256i hi = _mm512_extracti32x8_epi32(data, 1);
 #ifdef NPY_HAVE_AVX512BW
-    r.val[0] = _mm512_cvtepu8_epi16(_mm512_castsi512_si256(data));
-    r.val[1] = _mm512_cvtepu8_epi16(_mm512_extracti32x8_epi32(data, 1));
+    r.val[0] = _mm512_cvtepu8_epi16(lo);
+    r.val[1] = _mm512_cvtepu8_epi16(hi);
 #else
-    __m512i zero = npyv_zero_u8();
-    r.val[0] = _mm512_unpacklo_epi8(data, zero);
-    r.val[1] = _mm512_unpackhi_epi8(data, zero);
+    __m256i loelo = _mm256_cvtepu8_epi16(_mm256_castsi256_si128(lo));
+    __m256i loehi = _mm256_cvtepu8_epi16(_mm256_extracti128_si256(lo, 1));
+    __m256i hielo = _mm256_cvtepu8_epi16(_mm256_castsi256_si128(hi));
+    __m256i hiehi = _mm256_cvtepu8_epi16(_mm256_extracti128_si256(hi, 1));
+    r.val[0] = _mm512_inserti64x4(_mm512_castsi256_si512(loelo), loehi, 1);
+    r.val[1] = _mm512_inserti64x4(_mm512_castsi256_si512(hielo), hiehi, 1);
 #endif
     return r;
 }
@@ -69,13 +74,18 @@ NPY_FINLINE npyv_u16x2 npyv_expand_u16_u8(npyv_u8 data)
 NPY_FINLINE npyv_u32x2 npyv_expand_u32_u16(npyv_u16 data)
 {
     npyv_u32x2 r;
+    __m256i lo = _mm512_castsi512_si256(data);
+    __m256i hi = _mm512_extracti32x8_epi32(data, 1);
 #ifdef NPY_HAVE_AVX512BW
-    r.val[0] = _mm512_cvtepu16_epi32(_mm512_castsi512_si256(data));
-    r.val[1] = _mm512_cvtepu16_epi32(_mm512_extracti32x8_epi32(data, 1));
+    r.val[0] = _mm512_cvtepu16_epi32(lo);
+    r.val[1] = _mm512_cvtepu16_epi32(hi);
 #else
-    __m512i zero = npyv_zero_u16();
-    r.val[0] = _mm512_unpacklo_epi16(data, zero);
-    r.val[1] = _mm512_unpackhi_epi16(data, zero);
+    __m256i loelo = _mm256_cvtepu16_epi32(_mm256_castsi256_si128(lo));
+    __m256i loehi = _mm256_cvtepu16_epi32(_mm256_extracti128_si256(lo, 1));
+    __m256i hielo = _mm256_cvtepu16_epi32(_mm256_castsi256_si128(hi));
+    __m256i hiehi = _mm256_cvtepu16_epi32(_mm256_extracti128_si256(hi, 1));
+    r.val[0] = _mm512_inserti64x4(_mm512_castsi256_si512(loelo), loehi, 1);
+    r.val[1] = _mm512_inserti64x4(_mm512_castsi256_si512(hielo), hiehi, 1);
 #endif
     return r;
 }