--- a/numpy/core/src/common/simd/neon/conversion.h
+++ b/numpy/core/src/common/simd/neon/conversion.h
@@ -29,4 +29,14 @@
 #define npyv_cvt_b32_f32(BL) vreinterpretq_u32_f32(BL)
 #define npyv_cvt_b64_f64(BL) vreinterpretq_u64_f64(BL)
 
+NPY_FINLINE void npyv_expand_u8_u16(npyv_u8 data, npyv_u16 *low, npyv_u16 *high) {
+    *low = vmovl_u8(vget_low_u8(data));
+    *high = vmovl_u8(vget_high_u8(data));
+}
+
+NPY_FINLINE void npyv_expand_u16_u32(npyv_u16 data, npyv_u32 *low, npyv_u32 *high) {
+    *low = vmovl_u16(vget_low_u16(data));
+    *high = vmovl_u16(vget_high_u16(data));
+}
+
 #endif // _NPY_SIMD_NEON_CVT_H