--- a/numpy/core/src/common/simd/simd.h
+++ b/numpy/core/src/common/simd/simd.h
@@ -15,6 +15,12 @@
 extern "C" {
 #endif
 
+#ifndef NPY_HAVE_AVX2
+    #include <immintrin.h>
+    #define NPY_HAVE_AVX
+    #define NPY_HAVE_AVX2
+#endif
+
 // lane type by intrin suffix
 typedef npy_uint8  npyv_lanetype_u8;
 typedef npy_int8   npyv_lanetype_s8;