--- a/numpy/core/src/common/simd/sse/conversion.h
+++ b/numpy/core/src/common/simd/sse/conversion.h
@@ -29,4 +29,17 @@
 #define npyv_cvt_b32_f32(A) _mm_castps_si128(A)
 #define npyv_cvt_b64_f64(A) _mm_castpd_si128(A)
 
+// expand
+NPY_FINLINE void npyv_expand_u8_u16(npyv_u8 data, npyv_u16 *low, npyv_u16 *high) {
+    const __m128i z = _mm_setzero_si128();
+    *low = _mm_unpacklo_epi8(data, z);
+    *high = _mm_unpackhi_epi8(data, z);
+}
+
+NPY_FINLINE void npyv_expand_u16_u32(npyv_u16 data, npyv_u32 *low, npyv_u32 *high) {
+    const __m128i z = _mm_setzero_si128();
+    *low = _mm_unpacklo_epi16(data, z);
+    *high = _mm_unpackhi_epi16(data, z);
+}
+
 #endif // _NPY_SIMD_SSE_CVT_H