--- a/numpy/core/src/common/simd/vsx/conversion.h
+++ b/numpy/core/src/common/simd/vsx/conversion.h
@@ -29,16 +29,19 @@
 #define npyv_cvt_b32_f32(A) ((npyv_b32) A)
 #define npyv_cvt_b64_f64(A) ((npyv_b64) A)
 
-NPY_FINLINE void npyv_expand_u8_u16(npyv_u8 data, npyv_u16 *low, npyv_u16 *high) {
+NPY_FINLINE npyv_u16x2 npyv_expand_u8_u16(npyv_u8 data) {
+    npyv_u16x2 r;
     npyv_u8 zero = npyv_zero_u8();
-    *low = (npyv_u16)vec_mergel(data, zero);
-    *high = (npyv_u16)vec_mergeh(data, zero);
+    r.val[0] = (npyv_u16)vec_mergel(data, zero);
+    r.val[1] = (npyv_u16)vec_mergeh(data, zero);
+    return r;
 }
 
-NPY_FINLINE void npyv_expand_u16_u32(npyv_u16 data, npyv_u32 *low, npyv_u32 *high) {
+NPY_FINLINE npyv_u32x2 npyv_expand_u16_u32(npyv_u16 data) {
+    npyv_u32x2 r;
     npyv_u16 zero = npyv_zero_u16();
-    *low = (npyv_u32)vec_mergel(data, zero);
-    *high = (npyv_u32)vec_mergeh(data, zero);
+    r.val[0] = (npyv_u32)vec_mergel(data, zero);
+    r.val[1] = (npyv_u32)vec_mergeh(data, zero);
 }
 
 #endif // _NPY_SIMD_VSX_CVT_H