--- a/numpy/core/src/common/simd/vsx/conversion.h
+++ b/numpy/core/src/common/simd/vsx/conversion.h
@@ -45,6 +45,7 @@ NPY_FINLINE npyv_u32x2 npyv_expand_u32_u16(npyv_u16 data)
     npyv_u16 zero = npyv_zero_u16();
     r.val[0] = (npyv_u32)vec_mergel(data, zero);
     r.val[1] = (npyv_u32)vec_mergeh(data, zero);
+    return r;
 }
 
 // convert boolean vector to integer bitfield