… numpy/core/src/multiarray/item_selection.c
… numpy/core/tests/test_simd.py