## Introduction

`np.count_nonzero` is a common operation in database, information-retrieval, cryptographic and machine-learning applications. It's reported that [the same OpenCV function](https://github.com/opencv/opencv/blob/master/modules/core/src/count_non_zero.simd.hpp), which uses universal intrinsics technique,  is nearly **_25x_** [faster than the Numpy's function](https://docs.opencv.org/master/dc/d71/tutorial_py_optimization.html), The algorithm there is easy to migrate into current USIMD framework after some investigation. The performance increased 35% with avx2 instrument.

## Benchmark

Here is the ASV benchmark result.
<details>
<summary>AVX2 enabled</summary>
<pre><code>
· Creating environments
· Discovering benchmarks
·· Uninstalling from virtualenv-py3.7-Cython
·· Building 4337b254 <countnz> for virtualenv-py3.7-Cython
·· Installing 4337b254 <countnz> into virtualenv-py3.7-Cython
· Running 6 total benchmarks (2 commits * 1 environments * 3 benchmarks)
[  0.00%] · For numpy commit 7a505741 <master> (round 1/2):
[  0.00%] ·· Building for virtualenv-py3.7-Cython
[  0.00%] ·· Benchmarking virtualenv-py3.7-Cython
[  8.33%] ··· Running (bench_core.CountNonzero.time_count_nonzero--).
[ 16.67%] ··· Running (bench_core.CountNonzero.time_count_nonzero_axis--).
[ 25.00%] ··· Running (bench_core.CountNonzero.time_count_nonzero_multi_axis--).
[ 25.00%] · For numpy commit 4337b254 <countnz> (round 1/2):
[ 25.00%] ·· Building for virtualenv-py3.7-Cython
[ 25.00%] ·· Benchmarking virtualenv-py3.7-Cython
[ 33.33%] ··· Running (bench_core.CountNonzero.time_count_nonzero--).
[ 41.67%] ··· Running (bench_core.CountNonzero.time_count_nonzero_axis--).
[ 50.00%] ··· Running (bench_core.CountNonzero.time_count_nonzero_multi_axis--).
[ 50.00%] · For numpy commit 4337b254 <countnz> (round 2/2):
[ 50.00%] ·· Benchmarking virtualenv-py3.7-Cython
[ 58.33%] ··· bench_core.CountNonzero.time_count_nonzero                                                                            ok
[ 58.33%] ··· ========= ========= ============= ============= ============= =============
              --                                           dtype
              ------------------- -------------------------------------------------------
               numaxes     size        bool          int           str          object
              ========= ========= ============= ============= ============= =============
                  1        100      1.36±0.1µs   1.50±0.01µs   2.12±0.05µs   2.02±0.05µs
                  1       10000     1.99±0.3µs    32.1±0.6µs     85.9±4µs     77.2±0.6µs
                  1      1000000     41.2±2µs    2.99±0.01ms    9.55±0.3ms    8.15±0.1ms
                  2        100     1.28±0.01µs   1.74±0.03µs   2.93±0.06µs    2.75±0.1µs
                  2       10000    2.34±0.09µs    61.7±0.5µs     174±9µs       164±10µs
                  2      1000000     94.9±3µs    6.01±0.07ms     19.2±1ms     15.8±0.2ms
                  3        100     1.29±0.06µs    2.22±0.1µs   3.93±0.08µs   3.65±0.06µs
                  3       10000    2.81±0.07µs     91.6±1µs      257±8µs       259±8µs
                  3      1000000     141±5µs      9.14±0.1ms    28.5±0.4ms    23.6±0.6ms
              ========= ========= ============= ============= ============= =============

[ 66.67%] ··· bench_core.CountNonzero.time_count_nonzero_axis                                                                       ok
[ 66.67%] ··· ========= ========= ============= ============= ============ ============
              --                                          dtype
              ------------------- -----------------------------------------------------
               numaxes     size        bool          int          str         object
              ========= ========= ============= ============= ============ ============
                  1        100      8.57±0.3µs     11.5±1µs    11.0±0.1µs   10.7±0.5µs
                  1       10000      23.9±1µs      30.7±1µs     118±3µs      151±3µs
                  1      1000000   1.37±0.02ms   2.61±0.09ms    12.7±1ms    14.3±0.7ms
                  2        100      8.59±0.4µs     9.98±1µs    12.1±0.7µs    14.0±2µs
                  2       10000      39.0±2µs     49.5±0.7µs    246±5µs      290±10µs
                  2      1000000   2.76±0.04ms   4.87±0.09ms   24.2±0.7ms    28.0±1ms
                  3        100      8.58±0.5µs     10.2±1µs    13.6±0.5µs   13.5±0.4µs
                  3       10000      52.7±4µs      69.1±2µs     365±9µs      428±10µs
                  3      1000000   4.17±0.07ms    7.52±0.4ms   36.2±0.5ms   41.8±0.4ms
              ========= ========= ============= ============= ============ ============

[ 75.00%] ··· bench_core.CountNonzero.time_count_nonzero_multi_axis                                                                 ok
[ 75.00%] ··· ========= ========= ============= ============= ============ ============
              --                                          dtype
              ------------------- -----------------------------------------------------
               numaxes     size        bool          int          str         object
              ========= ========= ============= ============= ============ ============
                  1        100      8.79±0.7µs    9.58±0.3µs   11.9±0.9µs   11.0±0.3µs
                  1       10000     23.4±0.8µs    30.6±0.3µs    125±4µs      150±4µs
                  1      1000000   1.38±0.01ms   2.42±0.04ms   12.2±0.2ms   14.1±0.7ms
                  2        100      8.61±0.5µs    9.81±0.5µs   12.6±0.4µs    12.6±2µs
                  2       10000      38.1±1µs      57.0±4µs     245±20µs     294±10µs
                  2      1000000   2.80±0.06ms    4.84±0.1ms   25.1±0.3ms   27.7±0.2ms
                  3        100      8.75±0.3µs    9.95±0.3µs   13.5±0.6µs   13.8±0.4µs
                  3       10000     49.6±0.7µs     70.8±2µs     357±4µs      424±20µs
                  3      1000000   4.13±0.06ms    7.79±0.7ms   35.5±0.8ms   42.4±0.5ms
              ========= ========= ============= ============= ============ ============

[ 75.00%] · For numpy commit 7a505741 <master> (round 2/2):
[ 75.00%] ·· Building for virtualenv-py3.7-Cython
[ 75.00%] ·· Benchmarking virtualenv-py3.7-Cython
[ 83.33%] ··· bench_core.CountNonzero.time_count_nonzero                                                                            ok
[ 83.33%] ··· ========= ========= ============= ============= ============= =============
              --                                           dtype
              ------------------- -------------------------------------------------------
               numaxes     size        bool          int           str          object
              ========= ========= ============= ============= ============= =============
                  1        100     1.27±0.03µs   1.51±0.05µs    2.36±0.2µs   1.94±0.01µs
                  1       10000    2.13±0.04µs    32.5±0.7µs     85.3±2µs      79.1±1µs
                  1      1000000    63.2±0.8µs   2.99±0.03ms    9.28±0.2ms    7.93±0.3ms
                  2        100     1.27±0.02µs   1.79±0.06µs   2.94±0.07µs    2.84±0.1µs
                  2       10000    2.82±0.05µs    62.7±0.9µs     182±9µs       158±5µs
                  2      1000000     128±2µs     6.16±0.07ms    19.7±0.8ms    15.9±0.3ms
                  3        100     1.28±0.04µs   2.17±0.07µs   3.90±0.06µs   3.63±0.02µs
                  3       10000    3.42±0.06µs     91.2±2µs      263±10µs      236±3µs
                  3      1000000     197±8µs     9.05±0.06ms    28.0±0.2ms    23.7±0.2ms
              ========= ========= ============= ============= ============= =============

[ 91.67%] ··· bench_core.CountNonzero.time_count_nonzero_axis                                                                       ok
[ 91.67%] ··· ========= ========= ============= ============= ============= ============
              --                                          dtype
              ------------------- ------------------------------------------------------
               numaxes     size        bool          int           str         object
              ========= ========= ============= ============= ============= ============
                  1        100      8.26±0.2µs    9.51±0.8µs   10.6±0.07µs   11.1±0.6µs
                  1       10000     22.4±0.5µs     30.6±1µs      120±3µs      156±10µs
                  1      1000000   1.39±0.02ms   2.37±0.04ms    12.6±0.3ms   14.3±0.3ms
                  2        100      8.50±0.4µs    9.58±0.3µs    12.0±0.4µs   11.7±0.2µs
                  2       10000     38.3±0.8µs     50.9±3µs      234±4µs      295±20µs
                  2      1000000   2.75±0.02ms    4.97±0.2ms    24.7±0.6ms   28.2±0.4ms
                  3        100      8.66±0.4µs    9.55±0.1µs    13.6±0.5µs   13.5±0.3µs
                  3       10000      53.7±7µs      70.1±2µs      358±40µs     416±7µs
                  3      1000000    4.11±0.1ms    7.27±0.2ms     36.2±1ms     43.2±1ms
              ========= ========= ============= ============= ============= ============

[100.00%] ··· bench_core.CountNonzero.time_count_nonzero_multi_axis                                                                 ok
[100.00%] ··· ========= ========= ============= ============= ============ ============
              --                                          dtype
              ------------------- -----------------------------------------------------
               numaxes     size        bool          int          str         object
              ========= ========= ============= ============= ============ ============
                  1        100      8.64±0.4µs    9.57±0.3µs   11.4±0.3µs   10.8±0.3µs
                  1       10000      24.8±1µs      31.4±1µs     125±6µs      152±2µs
                  1      1000000    1.40±0.2ms   2.44±0.05ms   12.4±0.2ms   14.1±0.2ms
                  2        100      8.79±0.2µs    10.2±0.5µs   12.2±0.4µs    12.4±1µs
                  2       10000      36.8±1µs      52.8±4µs     247±10µs     292±20µs
                  2      1000000   2.80±0.04ms   5.00±0.07ms   24.7±0.7ms   28.0±0.7ms
                  3        100      9.24±0.3µs    10.1±0.3µs   13.9±0.4µs   14.0±0.4µs
                  3       10000     50.7±0.8µs     74.7±4µs     345±20µs     419±9µs
                  3      1000000    4.15±0.2ms    7.89±0.4ms    36.6±1ms     43.0±2ms
              ========= ========= ============= ============= ============ ============

       before           after         ratio
     [7a505741]       [4337b254]
     <master>         <countnz>
-     2.82±0.05µs      2.34±0.09µs     0.83  bench_core.CountNonzero.time_count_nonzero(2, 10000, <class 'bool'>)
-     3.42±0.06µs      2.81±0.07µs     0.82  bench_core.CountNonzero.time_count_nonzero(3, 10000, <class 'bool'>)
-         128±2µs         94.9±3µs     0.74  bench_core.CountNonzero.time_count_nonzero(2, 1000000, <class 'bool'>)
-         197±8µs          141±5µs     0.71  bench_core.CountNonzero.time_count_nonzero(3, 1000000, <class 'bool'>)
-      63.2±0.8µs         41.2±2µs     0.65  bench_core.CountNonzero.time_count_nonzero(1, 1000000, <class 'bool'>)

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
</code></pre>
</details>

<details>
<summary>SSE2 enabled</summary>
<pre><code>
· Creating environments
· Discovering benchmarks
·· Uninstalling from virtualenv-py3.7-Cython
·· Building d2f477c3 <countnz> for virtualenv-py3.7-Cython
·· Installing d2f477c3 <countnz> into virtualenv-py3.7-Cython
· Running 6 total benchmarks (2 commits * 1 environments * 3 benchmarks)
[  0.00%] · For numpy commit 7a505741 <master> (round 1/2):
[  0.00%] ·· Building for virtualenv-py3.7-Cython
[  0.00%] ·· Benchmarking virtualenv-py3.7-Cython
[  8.33%] ··· Running (bench_core.CountNonzero.time_count_nonzero--).
[ 16.67%] ··· Running (bench_core.CountNonzero.time_count_nonzero_axis--).
[ 25.00%] ··· Running (bench_core.CountNonzero.time_count_nonzero_multi_axis--).
[ 25.00%] · For numpy commit d2f477c3 <countnz> (round 1/2):
[ 25.00%] ·· Building for virtualenv-py3.7-Cython
[ 25.00%] ·· Benchmarking virtualenv-py3.7-Cython
[ 33.33%] ··· Running (bench_core.CountNonzero.time_count_nonzero--).
[ 41.67%] ··· Running (bench_core.CountNonzero.time_count_nonzero_axis--).
[ 50.00%] ··· Running (bench_core.CountNonzero.time_count_nonzero_multi_axis--).
[ 50.00%] · For numpy commit d2f477c3 <countnz> (round 2/2):
[ 50.00%] ·· Benchmarking virtualenv-py3.7-Cython
[ 58.33%] ··· bench_core.CountNonzero.time_count_nonzero                  ok
[ 58.33%] ··· ========= ========= ======== =============
               numaxes     size    dtype
              --------- --------- -------- -------------
                  1        100      bool    1.35±0.07μs
                  1        100      int     1.57±0.04μs
                  1        100      str     2.20±0.08μs
                  1        100     object    2.06±0.2μs
                  1       10000     bool     2.28±0.2μs
                  1       10000     int       32.2±1μs
                  1       10000     str      86.9±0.9μs
                  1       10000    object     77.7±1μs
                  1      1000000    bool      72.7±1μs
                  1      1000000    int     3.05±0.04ms
                  1      1000000    str      9.61±0.2ms
                  1      1000000   object    7.97±0.2ms
                  2        100      bool    1.36±0.03μs
                  2        100      int     1.82±0.03μs
                  2        100      str     2.97±0.09μs
                  2        100     object   2.83±0.08μs
                  2       10000     bool     3.12±0.2μs
                  2       10000     int       62.0±1μs
                  2       10000     str       173±10μs
                  2       10000    object     156±2μs
                  2      1000000    bool      146±5μs
                  2      1000000    int     6.10±0.07ms
                  2      1000000    str      19.0±0.6ms
                  2      1000000   object    16.0±0.4ms
                  3        100      bool    1.30±0.02μs
                  3        100      int      2.17±0.2μs
                  3        100      str     3.92±0.09μs
                  3        100     object    3.61±0.1μs
                  3       10000     bool     4.03±0.3μs
                  3       10000     int       95.5±2μs
                  3       10000     str       260±9μs
                  3       10000    object     234±5μs
                  3      1000000    bool      224±8μs
                  3      1000000    int      9.27±0.2ms
                  3      1000000    str      29.2±0.8ms
                  3      1000000   object    24.2±0.2ms
              ========= ========= ======== =============

[ 66.67%] ··· ...core.CountNonzero.time_count_nonzero_axis                ok
[ 66.67%] ··· ========= ========= ======== =============
               numaxes     size    dtype
              --------- --------- -------- -------------
                  1        100      bool     8.77±0.6μs
                  1        100      int      9.40±0.3μs
                  1        100      str      10.8±0.3μs
                  1        100     object    10.9±0.6μs
                  1       10000     bool     22.8±0.7μs
                  1       10000     int      31.7±0.9μs
                  1       10000     str       127±3μs
                  1       10000    object     148±3μs
                  1      1000000    bool    1.45±0.04ms
                  1      1000000    int      2.46±0.1ms
                  1      1000000    str      12.4±0.3ms
                  1      1000000   object    14.2±0.3ms
                  2        100      bool     8.67±0.2μs
                  2        100      int      10.0±0.5μs
                  2        100      str       13.5±2μs
                  2        100     object    13.2±0.6μs
                  2       10000     bool      41.4±2μs
                  2       10000     int       51.9±2μs
                  2       10000     str       242±10μs
                  2       10000    object     293±7μs
                  2      1000000    bool     2.83±0.1ms
                  2      1000000    int      5.09±0.1ms
                  2      1000000    str       25.5±1ms
                  2      1000000   object    27.9±0.4ms
                  3        100      bool     8.76±0.3μs
                  3        100      int      9.92±0.7μs
                  3        100      str      13.4±0.6μs
                  3        100     object     13.9±2μs
                  3       10000     bool      52.5±2μs
                  3       10000     int       71.3±1μs
                  3       10000     str       380±10μs
                  3       10000    object     432±20μs
                  3      1000000    bool     4.30±0.1ms
                  3      1000000    int      7.49±0.1ms
                  3      1000000    str       37.9±1ms
                  3      1000000   object    42.3±0.2ms
              ========= ========= ======== =============

[ 75.00%] ··· ...ountNonzero.time_count_nonzero_multi_axis                ok
[ 75.00%] ··· ========= ========= ======== =============
               numaxes     size    dtype
              --------- --------- -------- -------------
                  1        100      bool     8.93±0.3μs
                  1        100      int      9.64±0.3μs
                  1        100      str      11.3±0.4μs
                  1        100     object    11.1±0.3μs
                  1       10000     bool      23.6±1μs
                  1       10000     int       31.9±1μs
                  1       10000     str       130±10μs
                  1       10000    object     158±10μs
                  1      1000000    bool    1.50±0.07ms
                  1      1000000    int      2.50±0.1ms
                  1      1000000    str      12.3±0.3ms
                  1      1000000   object    14.4±0.7ms
                  2        100      bool     9.35±0.6μs
                  2        100      int      11.4±0.6μs
                  2        100      str       12.5±1μs
                  2        100     object    12.8±0.4μs
                  2       10000     bool      37.4±2μs
                  2       10000     int       53.3±2μs
                  2       10000     str       249±6μs
                  2       10000    object     298±8μs
                  2      1000000    bool     3.22±0.2ms
                  2      1000000    int      4.93±0.2ms
                  2      1000000    str      24.7±0.8ms
                  2      1000000   object    28.1±0.6ms
                  3        100      bool     9.28±0.2μs
                  3        100      int      10.4±0.4μs
                  3        100      str      13.6±0.7μs
                  3        100     object    15.6±0.7μs
                  3       10000     bool      51.9±1μs
                  3       10000     int       76.9±3μs
                  3       10000     str       386±30μs
                  3       10000    object     433±10μs
                  3      1000000    bool     4.85±0.3ms
                  3      1000000    int      7.55±0.3ms
                  3      1000000    str      36.9±0.8ms
                  3      1000000   object     43.5±1ms
              ========= ========= ======== =============

[ 75.00%] · For numpy commit 7a505741 <master> (round 2/2):
[ 75.00%] ·· Building for virtualenv-py3.7-Cython
[ 75.00%] ·· Benchmarking virtualenv-py3.7-Cython
…