@Qiyu8 2020-12-08
@seiko2plus when I do the `npyv_u8 vt = npyv_cmpeq_u8(a, b);` , The power machine has an error ```incompatible types when assigning to type ‘npyv_u8 {aka __vector(16) unsigned char}’ from type ‘__vector __bool char {aka __vector(16) __bool char}’```, Do you have any idea why this happened?