@seiko2plus 2020-12-08
@Qiyu8, because all comparison operations returns boolean vector data type, NPYV count on SIMD extensions
for type safety, so  it should be `npyv_b8 bt = npyv_cmpeq_u8(a, b);` otherwise it will break the build on `VSX` and `AVX512`