@eric-wieser 2020-12-08
I think something like this would avoid overflow issues:
```C
static NPY_INLINE npyv_u8
count_zero_bytes_u8(const npy_uint8 *d, const npy_uint8 *end)
{
    const npyv_u8 vone = npyv_setall_u8(1);
    const npyv_u8 vzero = npyv_setall_u8(0);

    npy_intp n = 0;
    npyv_u8 vsum8 = npyv_zero_u8();
    while (*d < end && n <= 0xFE) {
        npyv_b8 vt = npyv_cmpeq_u8(npyv_load_u8(d), vzero);
        vt = npyv_and_u8(vt, vone);
        vsum8 = npyv_add_u8(vsum8, vt);
        d += npyv_nlanes_u8;
        n++;
    }
    return vsum8;
}

static NPY_INLINE npyv_u16
count_zero_bytes_u16(const npy_uint8 **d, const npy_uint8 *end)
{
    npyv_u16 vsum16 = npyv_zero_u16();
    npy_intp n = 0;
    while (*d < end && n <= 0xFF00) {
        npyv_u8 vsum8 = count_zero_bytes_u8(d, end);
        npyv_u16 part1, part2;
        npyv_expand_u8_u16(vsum8, &part1, &part2);
        vsum16 = npyv_add_u16(vsum16, npyv_add_u16(part1, part2));
        n += 0xFF;
    }
    return vsum16;
}

static NPY_INLINE npyv_u32
count_zero_bytes_u32(const npy_uint8 **d, const npy_uint8 *end)
{
    npyv_u32 vsum32 = npyv_zero_u32();
    npy_intp n = 0;
    while (*d < end && n <= 0xFFFF0000) {
        npyv_u8 vsum16 = count_zero_bytes_u16(d, end);
        npyv_u32 part1, part2;
        npyv_expand_u16_u32(vsum16, &part1, &part2);
        vsum32 = npyv_add_u32(vsum32, npyv_add_u32(part1, part2));
        n += 0xFFFF;
    }
    return vsum32;
}


static NPY_INLINE npy_intp
count_nonzero_bytes(const npy_uint8 *d, npy_uintp unrollx)
{
    npy_intp zero_count = 0;
    const npy_uint8 *end = d + unrollx;
    while (*d < end) {
        npyv_u32 vsum32 = count_zero_bytes_u32(d, end);
        zero_count += npyv_sum_u32(vsum32);
    }
    return unrollx - zero_count;
}
```