@Qiyu8 2020-12-09
@eric-wieser The over-flow preventing code you presented looks more elegant, I will try and give a benchmark result. thanks.