@Qiyu8 2020-12-09
@eric-wieser Here is the benchmark result of the modular code.
```
       before           after         ratio
     [7a505741]       [dfbb3f76]
     <master>         <countnz>
-         133±3μs         89.9±5μs     0.67  bench_core.CountNonzero.time_count_nonzero(2, 1000000, <class 'bool'>)
-         201±8μs          135±3μs     0.67  bench_core.CountNonzero.time_count_nonzero(3, 1000000, <class 'bool'>)
-        64.5±2μs         43.1±1μs     0.67  bench_core.CountNonzero.time_count_nonzero(1, 1000000, <class 'bool'>)

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```
The performance remains good, so I will take that code.