@Qiyu8 2020-12-09
`_mm512_cvtepu8_epi16` is only available in `AVX512BW`, @seiko2plus Any idea in implement the expansion in other platform like `AVX512F`?