@Qiyu8 2020-12-15
The test case of new intrinsics has added, so the `codecov/patch` result is inaccurate here.