@seiko2plus 2020-12-15
@Qiyu8, codecov builders sometimes run on x86 machines with no avx512 support